pyproject.toml
[project]
name = "azure-mcp-server"
version = "1.0.0"
description = "MCP server for Azure resource management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "azure-mgmt-resource",
    "azure-mgmt-compute",
    "azure-mgmt-network",
    "azure-mgmt-sql",
    "azure-mgmt-containerregistry",
    "azure-mgmt-web",
    "azure-mgmt-dns",
    "azure-identity",
    "python-dotenv",
    "pyproject-toml>=0.1.0",
    "paramiko>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
azure-mcp-server = "azure_mcp_server.main:run_server"

[tool.setuptools]
packages = ["azure_mcp_server"]
package-dir = {"" = "src"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src/azure_mcp_server"]
omit = ["*/__pycache__/*", "*/tests/*", "*/.venv/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]
